parcel_service/requirements.txt
fastapi
uvicorn
python-jose
pydantic
sqlalchemy
psycopg2-binary
